+++ src/unpoly/fragment.js
@@ -166,18 +166,10 @@ up.fragment = (function() {
     - If focus was lost with the old fragment, re-focus a [similar](/target-derivation) element.
     - If focus was lost with the old fragment, focus the new fragment.
 
-  @param {boolean} [config.runScripts=false]
+  @param {boolean|Function(ScriptElement): boolean} [config.runScripts=false]
     Whether to load or execute `<script>` tags in updated fragments.
 
-    Note that the `<body>` element is a default
-    [main target](/main). If you are including your global application scripts
-    at the end of your `<body>` for performance reasons, swapping the `<body>` will re-execute
-    these scripts. In that case you can [configure a different main target](/up.fragment.config#config.mainTargets)
-    or [move your scripts to the `<head>` with a `[defer]` attribute](https://makandracards.com/makandra/504104-you-should-probably-load-your-javascript-with-script-defer),
-    which is even better for performance.
-
-    Scripts in the `<head>` are never load or executed, even with `up.fragment.config.runScripts = true`.
-    See [Handling new asset versions](/handling-asset-changes#handling-changed-assets).
+    See [Running inline `<script>` tags](/legacy-scripts#running-inline-script-tags) for details.
 
   @param {boolean|Function(up.Response): boolean} [config.autoRevalidate]
     Whether to reload a fragment after it was rendered from a cached response with `{ revalidate: 'auto' }`.